fix[notask]: ocr-ggml clang-tidy bucket 3 - correctness
Resolves 43 errors across 11 clang-tidy check categories (build +
unit tests pass).

Per-check breakdown:
- bugprone-throwing-static-initialization             10 (NOLINT
  on the namespace-scope `const cv::Scalar`, `const std::string`,
  and `const std::map` initializers: VOCAB, DOCTR_DET_MEAN/STD,
  DOCTR_RECO_MEAN/STD, DEFAULT_MEAN/VARIANCE,
  RECOGNITION_MODELS, LANG_TO_CHARS_MAP, g_opencvInit;
  refactoring to function-local statics would force every consumer
  through a getter for no real-world benefit)
- bugprone-implicit-widening-of-multiplication-result 11 (cast the
  first multiplicand to size_t before the multiplication so the
  whole expression evaluates in size_t domain)
- bugprone-misplaced-widening-cast                     7 (move the
  size_t cast onto a single operand rather than the int*int result
  to actually widen before multiplication)
- clang-diagnostic-deprecated-declarations             3 (NOLINT
  on `std::wstring_convert<std::codecvt_utf8<char32_t>>` which has
  no drop-in C++20 replacement; ICU/iconv swap is a separate
  refactor)
- clang-analyzer-deadcode.DeadStores                   3 (removed
  the unused `addFcWeightFp32` and `addFcBiasFp32` lambdas in
  MobileNetGraph.cpp; NOLINT on `uploadTensorBytes` which is
  retained as documentation of the classifier-bytes upload path)
- clang-diagnostic-deprecated-enum-enum-conversion     2 (cast
  GGML_SCALE_MODE_BILINEAR and GGML_SCALE_FLAG_ALIGN_CORNERS to
  uint32_t before the bitwise OR — ggml's C API mixes enum types)
- bugprone-narrowing-conversions                       2 (widened
  `int idx` to `int64_t` in MobileNetGraph.cpp::cloneRaw to match
  gguf_find_tensor's return type; cast hardware_concurrency() to
  int explicitly in StepDoctrDetectionGGML.cpp)
- readability-redundant-casting                        1 (dropped
  redundant `static_cast<int>(c)` inside `pixel[c]`)
- readability-isolate-declaration                      1 (split
  `std::vector<float> w_folded, b_folded;` in crnn_weights.cpp)
- performance-unnecessary-copy-initialization          1 (changed
  `const cv::Vec3f pixel = image.at<cv::Vec3f>(y, x)` to a const
  reference, then to `const auto&` so use-auto is also happy)
- bugprone-switch-missing-default-case                 1 (added
  `default: break;` to the FPN feature-tap switch in
  MobileNetGraph.cpp)

No behavioural changes. The narrowing fix in cloneRaw is a strict
widening (int -> int64_t), and the ggml enum bitwise OR semantics
are preserved by the explicit uint32_t cast.

+++ packages/ocr-ggml/addon/src/model-interface/doctr/StepDoctrDetectionGGML.cpp
@@ -22,8 +22,10 @@ namespace {
 
 // DocTR detection normalisation constants (from HuggingFace db_resnet50
 // config).
+// NOLINTBEGIN(bugprone-throwing-static-initialization)
 const cv::Scalar DOCTR_DET_MEAN(0.798, 0.785, 0.772);
 const cv::Scalar DOCTR_DET_STD(0.264, 0.2749, 0.287);
+// NOLINTEND(bugprone-throwing-static-initialization)
 constexpr double PIXEL_MAX = 255.0;
 
 // Mean probability inside a bounding rectangle (assume_straight_pages=True).
@@ -55,7 +57,7 @@ StepDoctrDetectionGGML::StepDoctrDetectionGGML(
     ggml_backend_cpu_set_n_threads(cpuBackend, nThreads);
   } else if (nThreads == 0) {
     ggml_backend_cpu_set_n_threads(
-        cpuBackend, std::thread::hardware_concurrency());
+        cpuBackend, static_cast<int>(std::thread::hardware_concurrency()));
   }
   backends_.push_back(cpuBackend);
